[FAB-8705] Remove samplesyscc package
The sample system chaincode was used as part of a system chaincode test
so moving it closer to where it's used.

That said, the test that uses it has been skipped for a long time and it
no longer passes when enabled.
